requirements.txt
kiteconnect==5.0.0
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
